apis/twitter/requirements.txt
# API Twitter - Dependências
# Python 3.10+

# Framework Web
flask>=3.0.0
gunicorn>=21.0.0

# Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-secret-manager>=2.16.0
google-auth>=2.22.0

# HTTP Requests
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.25.0

# Date/Time
pytz>=2023.3
python-dateutil>=2.8.2

# Utilities
cachetools>=5.3.0
